[BUGFIX] Fix loading of resources with spaces in filenames
Encode URI loaded from S3 to prevent problems with browsers that
don't fully support spaces in URIs.

* File location changed
